+++ Database/Updates/World/2018-01-07-00-inventory-shade.sql
@@ -1,0 +1,5 @@
+USE `ace_world`;
+
+ALTER TABLE `ace_object_inventory` 
+ADD COLUMN `shade` FLOAT NOT NULL DEFAULT '1' AFTER `palette`,
+ADD COLUMN `tryToBond` TINYINT(4) NULL DEFAULT '0' AFTER `shade`;